requirements.txt
covalent-api-sdk~=1.0.0
flask~=3.0.3
requests~=2.32.3